fix: 리뷰작성, 참여 안되는 에러 수정

+++ src/app/api/moims/[id]/review/[reviewId]/route.ts
@@ -1,3 +1,4 @@
+import { TMe } from '@/types/auth/auth.type';
 import { TReviewInput, TReviews } from '@/types/supabase/supabase-custom.type';
 import { mapMoimsToClient } from '@/utils/common/mapMoims';
 import { createClient } from '@/utils/supabase/server';
@@ -29,13 +30,30 @@ export async function PUT(
   }
 
   const {
+    data: foundUser,
+    error: foundUserError,
+  }: { data: TMe | null; error: PostgrestError | null } = await supabase
+    .from('users')
+    .select('*')
+    .eq('email', user.email)
+    .single();
+
+  if (foundUserError) {
+    return NextResponse.json({ message: foundUserError?.message }, { status: 401 });
+  }
+
+  if (!foundUser) {
+    return NextResponse.json({ message: '사용자 정보가 없어요' }, { status: 401 });
+  }
+
+  const {
     data: updatedReview,
     error: updatedReviewError,
   }: { data: TReviews | null; error: PostgrestError | null } = await supabase
     .from('reviews')
     .update({ review, rate, updated_at: new Date().toISOString() })
     .eq('id', reviewId)
-    .eq('user_uuid', user.id)
+    .eq('user_uuid', foundUser.id)
     .eq('moim_uuid', id)
     .select()
     .single();
@@ -90,12 +108,29 @@ export async function DELETE(
     return NextResponse.json({ message: '유저가 없습니다' }, { status: 401 });
   }
 
+  const {
+    data: foundUser,
+    error: foundUserError,
+  }: { data: TMe | null; error: PostgrestError | null } = await supabase
+    .from('users')
+    .select('*')
+    .eq('email', user.email)
+    .single();
+
+  if (foundUserError) {
+    return NextResponse.json({ message: foundUserError?.message }, { status: 401 });
+  }
+
+  if (!foundUser) {
+    return NextResponse.json({ message: '사용자 정보가 없어요' }, { status: 401 });
+  }
+
   const { data: deletedReview, error: deletedReviewError } = await supabase
     .from('reviews')
     .delete()
     .eq('id', reviewId)
     .eq('moim_uuid', id)
-    .eq('user_uuid', user.id)
+    .eq('user_uuid', foundUser.id)
     .select()
     .single();
